extracts design_system
