+++ Cargo.toml
@@ -4,6 +4,8 @@ members = [
     "anna_app",
     "anna_chat",
     "anna_profile",
+    "anna_design_system",
 ]
+
 exclude = [
 ]
