+++ anna_app/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "anna_app"
 version = "0.0.1"
-authors = ["huhn <huhn@einfachiota.de"]
+authors = ["huhn <huhn@einfachiota.de>"]
 edition = "2018"
 
 [lib]
@@ -40,6 +40,7 @@ anna_home = { path = "../anna_home" }
 anna_chat = { path = "../anna_chat" }
 anna_profile = { path = "../anna_profile" }
 anna_shop = { path = "../anna_shop" }
+anna_design_system = { path = "../anna_design_system" }
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3"
