+++ anna_app/src/components/layout.rs
@@ -1,5 +1,3 @@
-use crate::design_system::Footer;
-use crate::design_system::Header;
 use css_in_rust::Style;
 use yew::prelude::*;
 use yew::{html, Classes, Component, ComponentLink, Html, Properties, ShouldRender};
@@ -60,11 +58,9 @@ impl Component for Layout {
     fn view(&self) -> Html {
         html! {
             <div class=Classes::from(self.props.class.to_string()).extend(self.style.to_string())>
-                <Header />
                 <div class="content">
                     { self.props.children.clone() }
                 </div>
-                <Footer />
             </div>
         }
     }
