--- anna_app/src/design_system/atoms/logo/style.scss
@@ -1,3 +1,0 @@
-& {
-  margin-right: 50px;
-}
