+++ anna_app/src/pages/scanner.rs
@@ -1,4 +1,3 @@
-use crate::design_system::Button;
 use css_in_rust::Style;
 use yew::{html, Component, ComponentLink, Html, ShouldRender};
