+++ anna_design_system/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "anna_design_system"
+version = "0.0.1"
+authors = ["r3p0x <svreiser@gmail.com>", "huhn <huhn@einfachiota.de>"]
+edition = "2018"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+yew = { version = "^0.17", features = ["web_sys"] }
+css-in-rust = { version = "0.4.1", features = ["yew_integration"] }
