+++ anna_design_system/src/atoms/menu/style.scss
