+++ anna_design_system/src/atoms/menu_item/mod.rs
@@ -1,0 +1,3 @@
+pub mod menu_item;
+
+pub use menu_item::MenuItem;
