+++ anna_design_system/src/atoms/mod.rs
@@ -4,5 +4,8 @@ pub use logo::logo::Logo;
 pub mod menu;
 pub use menu::menu::Menu;
 
+pub mod menu_item;
+pub use menu_item::menu_item::MenuItem;
+
 pub mod page;
 pub use page::page::Page;
