+++ anna_design_system/src/lib.rs
@@ -1,3 +1,5 @@
+#![recursion_limit = "1024"]
+
 pub mod atoms;
 pub mod button;
 pub mod footer;
