+++ anna_shopping_cart/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "anna_shopping_cart"
 version = "0.0.1"
-authors = ["huhn <huhn@einfachiota.de"]
+authors = ["huhn <huhn@einfachiota.de>"]
 edition = "2018"
